separate out traits from impl
