+++ src/game.rs
@@ -1,4 +1,6 @@
 use crate::generator::debug::*;
+use crate::generator::*;
+
 use crate::hex_grid::*;
 use crate::piece::*;
 use crate::uhp::*;
