+++ src/generator/mod.rs
@@ -1,6 +1,8 @@
 pub mod change;
 pub mod debug;
 pub mod mini;
+pub mod generator;
 
 use super::*;
 use change::*;
+pub use generator::*;
