+++ src/testing_utils/positions.rs
@@ -1,4 +1,5 @@
 use crate::generator::debug::*;
+use crate::generator::generator::*;
 use crate::hex_grid::{HexGrid, HexLocation};
 use crate::piece::*;
 use crate::uhp::GameType;
